Log improvements - WIP: - move account logs (functional & security) in log plugin
