+++ src/main/core/Resources/config/services/controller.yml
@@ -329,28 +329,6 @@ services:
             - '@Claroline\CoreBundle\Manager\Workspace\EvaluationManager'
             - '@Claroline\AppBundle\Event\StrictDispatcher'
 
-    Claroline\CoreBundle\Controller\APINew\Log\SecurityLogController:
-        parent: Claroline\AppBundle\Controller\AbstractSecurityController
-        public: true
-        arguments:
-            - '@Claroline\AppBundle\API\FinderProvider'
-            - '@security.authorization_checker'
-            - '@security.token_storage'
-
-    Claroline\CoreBundle\Controller\APINew\Log\MessageLogController:
-        parent: Claroline\AppBundle\Controller\AbstractSecurityController
-        public: true
-        arguments:
-            - '@Claroline\AppBundle\API\FinderProvider'
-
-    Claroline\CoreBundle\Controller\APINew\Log\FunctionalLogController:
-        parent: Claroline\AppBundle\Controller\AbstractSecurityController
-        public: true
-        arguments:
-            - '@Claroline\AppBundle\API\FinderProvider'
-            - '@security.authorization_checker'
-            - '@security.token_storage'
-
     Claroline\CoreBundle\Controller\APINew\Location\LocationController:
         parent: Claroline\AppBundle\Controller\AbstractCrudController
         public: true
