+++ src/main/core/Resources/config/services/subscriber.yml
@@ -1,0 +1,13 @@
+services:
+    Claroline\CoreBundle\Subscriber\Crud\Planning\AbstractPlannedSubscriber:
+        abstract: true
+        calls:
+            - method: setTokenStorage
+              arguments:
+                  - '@security.token_storage'
+            - method: setObjectManager
+              arguments:
+                  - '@Claroline\AppBundle\Persistence\ObjectManager'
+            - method: setPlanningManager
+              arguments:
+                  - '@Claroline\CoreBundle\Manager\PlanningManager'
