--- src/main/core/Resources/modules/account/functional/store/index.js
@@ -1,7 +1,0 @@
-import {reducer} from '#/main/core/account/functional/store/reducer'
-import {selectors} from '#/main/core/account/functional/store/selectors'
-
-export {
-  reducer,
-  selectors
-}
