--- src/main/core/Resources/modules/account/security/store/index.js
@@ -1,7 +1,0 @@
-import {reducer} from '#/main/core/account/security/store/reducer'
-import {selectors} from '#/main/core/account/security/store/selectors'
-
-export {
-  reducer,
-  selectors
-}
