+++ src/main/log/Resources/config/services/subscriber.yml
@@ -1,17 +1,4 @@
 services:
-    Claroline\CoreBundle\Subscriber\Crud\Planning\AbstractPlannedSubscriber:
-        abstract: true
-        calls:
-            - method: setTokenStorage
-              arguments:
-                  - '@security.token_storage'
-            - method: setObjectManager
-              arguments:
-                  - '@Claroline\AppBundle\Persistence\ObjectManager'
-            - method: setPlanningManager
-              arguments:
-                  - '@Claroline\CoreBundle\Manager\PlanningManager'
-
     Claroline\LogBundle\Subscriber\SecurityLogSubscriber:
         arguments:
             - '@doctrine.orm.entity_manager'
