+++ src/main/log/Resources/modules/account/functional/components/list.jsx
