+++ src/main/log/Resources/modules/account/functional/containers/main.jsx
@@ -3,8 +3,8 @@ import {connect} from 'react-redux'
 import {withReducer} from '#/main/app/store/components/withReducer'
 import {selectors as securitySelectors} from '#/main/app/security/store'
 
-import {reducer, selectors} from '#/main/core/account/functional/store'
-import {FunctionalMain as FunctionalMainComponent}  from '#/main/core/account/functional/components/main'
+import {reducer, selectors} from '#/main/log/account/functional/store'
+import {FunctionalMain as FunctionalMainComponent}  from '#/main/log/account/functional/components/main'
 
 const FunctionalMain = withReducer(selectors.STORE_NAME, reducer)(
   connect(
