+++ src/main/log/Resources/modules/account/functional/store/index.js
@@ -1,0 +1,7 @@
+import {reducer} from '#/main/log/account/functional/store/reducer'
+import {selectors} from '#/main/log/account/functional/store/selectors'
+
+export {
+  reducer,
+  selectors
+}
