+++ src/main/log/Resources/modules/account/security/containers/main.jsx
@@ -3,8 +3,8 @@ import {connect} from 'react-redux'
 import {withReducer} from '#/main/app/store/components/withReducer'
 import {selectors as securitySelectors} from '#/main/app/security/store'
 
-import {reducer, selectors} from '#/main/core/account/security/store'
-import {SecurityMain as SecurityMainComponent}  from '#/main/core/account/security/components/main'
+import {reducer, selectors} from '#/main/log/account/security/store'
+import {SecurityMain as SecurityMainComponent}  from '#/main/log/account/security/components/main'
 
 const SecurityMain = withReducer(selectors.STORE_NAME, reducer)(
   connect(
