+++ src/main/log/Resources/modules/account/security/store/index.js
@@ -1,0 +1,7 @@
+import {reducer} from '#/main/log/account/security/store/reducer'
+import {selectors} from '#/main/log/account/security/store/selectors'
+
+export {
+  reducer,
+  selectors
+}
