+++ src/main/log/Resources/modules/account/security/store/reducer.js
@@ -1,6 +1,6 @@
 import {makeListReducer} from '#/main/app/content/list/store'
 
-import {selectors} from '#/main/core/account/functional/store/selectors'
+import {selectors} from '#/main/log/account/security/store/selectors'
 
 const reducer = makeListReducer(selectors.STORE_NAME, {
   sortBy: {property: 'date', direction: -1}
